refactor and update spacing
