+++ apps/src/codebridge/Workspace/index.tsx
@@ -1,8 +1,8 @@
 import {useCodebridgeContext} from '@codebridge/codebridgeContext';
+import ToggleFileBrowserButton from '@codebridge/components/ToggleFileBrowserButton';
 import {Editor} from '@codebridge/Editor';
 import {FileBrowser} from '@codebridge/FileBrowser';
 import {FileTabs} from '@codebridge/FileTabs';
-import ToggleFileBrowserButton from '@codebridge/ToggleFileBrowserButton';
 import classnames from 'classnames';
 import React from 'react';
