+++ apps/src/codebridge/components/ToggleFileBrowserButton.tsx
@@ -1,10 +1,12 @@
+import classNames from 'classnames';
 import React, {useCallback} from 'react';
 
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import Button from '@cdo/apps/componentLibrary/button';
 
-import {useCodebridgeContext} from './codebridgeContext';
+import {useCodebridgeContext} from '../codebridgeContext';
 
+import moduleStyles from './toggle-file-browser-button.module.scss';
 import darkModeStyles from '@codebridge/styles/darkMode.module.scss';
 
 /*
@@ -40,7 +42,10 @@ const ToggleFileBrowserButton: React.FunctionComponent = () => {
         ariaLabel={codebridgeI18n.toggleFileBrowser()}
         size={'xs'}
         type={'tertiary'}
-        className={darkModeStyles.iconOnlyTertiaryButton}
+        className={classNames(
+          darkModeStyles.iconOnlyTertiaryButton,
+          moduleStyles.button
+        )}
       />
     </span>
   );
