+++ apps/src/codebridge/components/toggle-file-browser-button.module.scss
@@ -1,0 +1,3 @@
+.button {
+  margin: 4px 8px;
+}
